Add API for Stock Buying function
Add BuyController API
Change Column Name of Investment to StartingInvestment in Account Table
of Database
Delete Column AvgBuyPrice in Transaction Table of Database
Change Data Type from Double to Decimal in Portfolio, Stock, Transaction
& Account Table

+++ StockDataServer/Controllers/PortfolioController.cs
@@ -61,7 +61,7 @@ namespace StockDataServer.Controllers
         }
 
         [HttpPost]
-        public bool InsertNewPortfolio(string username, string ticker, double cost,
+        public bool InsertNewPortfolio(string username, string ticker, decimal cost,
                                         int numStocks)
         {
             try
