+++ StockDataServer/Models/PortfolioTabModel.cs
@@ -9,12 +9,12 @@ namespace StockDataServer.Models
     {
         public string Ticker { get; set; }
         public string EquityName { get; set; }
-        public double Price { get; set; }
-        public double Cost { get; set; }
-        public double GainLossMoney { get; set; }
+        public decimal Price { get; set; }
+        public decimal Cost { get; set; }
+        public decimal GainLossMoney { get; set; }
         public long NumStocks { get; set; }
-        public double ChangeMoney { get; set; }
-        public double Value { get; set; }
-        public double ChangePercent { get; set; }
+        public decimal ChangeMoney { get; set; }
+        public decimal Value { get; set; }
+        public decimal ChangePercent { get; set; }
     }
 }
